destructured postscream further and managed to have no more distinction between mobile and desktop
